Share types used for testing JSON
