+++ test/Recore.Text.Json.Serialization.Converters/OptionalConverterTests.cs
@@ -6,12 +6,6 @@ namespace Recore.Text.Json.Serialization.Converters.Tests
 {
     public class OptionalConverterTests
     {
-        class Person
-        {
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
         [Fact]
         public void ToJson()
         {
